:tophat: Implement Imgur Api
Changes:
- Implement imgur "all of" search
- Refactor reddit search
- Use lodash
- Implement gif model
- Address technical debt including dockerfile and pm2
- Add github repo urls to package.json
- Revise package.json description
- Rempve scripts and use npm commands defined in package.json instead
- Remove unnecessary middleware
- Update readme
