+++ .dockerignore
@@ -2,9 +2,9 @@
 .gitignore
 .jshintrc
 .dockerignore
+.DS_Store
 Dockerfile
 docker-compose.yml
 README.md
 node_modules
 LICENSE
-script/development
