--- script/development/stop
@@ -1,9 +1,0 @@
-#!/bin/bash
-set -e
-
-GREEN="\033[0;32m"
-DEFAULT="\033[0m"
-
-echo -e "${GREEN}Stopping app${DEFAULT}"
-$(npm bin)/pm2 stop GIFts
-$(npm bin)/pm2 delete GIFts
