+++ server/index.js
@@ -3,13 +3,11 @@ import express from 'express';
 import cors from 'cors';
 import bodyParser from 'body-parser';
 import compression from 'compression';
-import middleware from './middleware';
 import api from './api';
 
 var app = express();
 app.server = http.createServer(app);
 
-// 3rd party middleware
 app.use(cors({
 	exposedHeaders: ['Link']
 }));
@@ -24,10 +22,6 @@ app.use(bodyParser.urlencoded({
 
 app.use(compression());
 
-// internal middleware
-app.use(middleware());
-
-// api router
 app.use('/api', api());
 
 app.server.listen(process.env.PORT || 8080);
