--- server/middleware/index.js
@@ -1,9 +1,0 @@
-import { Router } from 'express';
-
-export default function() {
-	var routes = Router();
-
-	// add middleware here
-
-	return routes;
-}
